added mvc-pattern, made a special directory for js css, added setup for migrations
