+++ .docker/nginx/default.conf
@@ -10,12 +10,10 @@ server {
         try_files $uri $uri/ /index.php?q=$uri&$args;
     }
 
-    location = /auth.php {
-        return 403;  
-    }
-
-    location = /register.php {
-        return 403;  
+    location /register.php {
+        if ($request_method = GET) {
+            return 403;  # Запретить доступ
+        }
     }
 
     location ~ \.php$ {
