+++ composer.json
@@ -10,7 +10,9 @@
             "Entities\\": "www/src/Entities/",
             "Repository\\": "www/src/Repository/",
             "Factory\\": "www/src/Factory/",
-            "Controllers\\": "www/src/Controllers/"
+            "Controllers\\": "www/src/Controllers/",
+            "Views\\" : "www/src/Views/",
+            "MiddleWares\\" : "www/src/MiddleWares"
         }
     }
 }
