+++ www/assets/png/logout.png
